fix(ingest): prevent malicious SEP-41 token from wedging ledger persistence
  SEP-41 defines decimals() -> u32, but contract_tokens.decimals was
  SMALLINT with the insert cast $8::smallint[]. A contract returning
  a decimals value above 32767 would fail pgx encoding, roll back
  the atomic ledger-persistence transaction, and prevent the ingest
  cursor from advancing past the affected ledger. The same wedge
  class also exists for name()/symbol() strings that exceed Postgres
  TEXT constraints (NUL byte, invalid UTF-8) or grow to unreasonable
  sizes. Reported via HackerOne.

  Widen contract_tokens.decimals to INTEGER (the service is not yet
  in production, so the existing migration is edited in place rather
  than adding a new one). Cap decimals at 70, name at 128 bytes,
  symbol at 32 bytes, and reject non-UTF-8 or NUL-containing strings
  in fetchMetadata. Violations fall through the existing warning path
  so the contract persists with sentinel values (decimals=0, name="",
  symbol="") instead of rolling back the ledger.

+++ internal/db/migrations/2025-06-13.0-contract_tokens.sql
@@ -10,7 +10,7 @@ CREATE TABLE contract_tokens (
     issuer TEXT NULL,
     name TEXT NULL,
     symbol TEXT NULL,
-    decimals SMALLINT NOT NULL,
+    decimals INTEGER NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
